+++ src/WebRoutes.tsx
@@ -32,6 +32,12 @@ import AddNewTask from "./pages/task/newTask/AddNewTask";
 import EditNewTask from "./pages/task/newTask/EditNewTask";
 import IndexInprogressTask from "./pages/task/improgressTask/IndexInprogressTask";
 import EditInprogressTask from "./pages/task/improgressTask/EditInprogressTask";
+import EditCancel from "./pages/task/todayTask/DetailEdit/EditCancel";
+import EditWaitStart from "./pages/task/todayTask/DetailEdit/EditWaitStart";
+import EditInProgress from "./pages/task/todayTask/DetailEdit/EditInProgress";
+import IndexTodayTask from "./pages/task/todayTask/IndexTaskToday";
+
+
 
 const WebRoutes: React.FC<any> = () => {
   return (
@@ -73,6 +79,10 @@ const WebRoutes: React.FC<any> = () => {
             path="/EditInprogressTask/:id"
             element={<EditInprogressTask />}
           />
+          <Route path="/IndexTodayTask" element={<IndexTodayTask />} />
+          <Route path="/EditWaitStart" element={<EditWaitStart/>} />
+          <Route path="/EditInProgress" element={<EditInProgress/>} />
+          <Route path="/EditCancel" element={<EditCancel/>} />
           <Route path="*" element={<PageNotFound />} />
         </Route>
         <Route element={<PublicRoute />}>